feat : 인수 테스트

+++ src/main/java/nextstep/subway/applicaion/dto/PathResponse.java
@@ -8,26 +8,26 @@ import java.util.stream.Collectors;
 public class PathResponse {
 
     private List<StationResponse> stations;
-    private double distance;
+    private int distance;
 
     public PathResponse() {
     }
 
-    public PathResponse(List<StationResponse> stations, double distance) {
+    public PathResponse(List<StationResponse> stations, int distance) {
         this.stations = stations;
         this.distance = distance;
     }
 
-    public static PathResponse of(List<Station> shortestPath,double pathWeight) {
+    public static PathResponse of(List<Station> shortestPath, int pathWeight) {
         List<StationResponse> stationResponses = shortestPath.stream().map(StationResponse::of).collect(Collectors.toList());
-        return new PathResponse(stationResponses,pathWeight);
+        return new PathResponse(stationResponses, pathWeight);
     }
 
     public List<StationResponse> getStations() {
         return stations;
     }
 
-    public double getDistance() {
+    public int getDistance() {
         return distance;
     }
 }
